+++ analysis/src/analysis.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use tokio_util::sync::CancellationToken;
 
 use crate::analyzer::core_analyzer::Analyzer;
-use crate::shared_utils::computing_device_type::ComputingDevice;
+use crate::shared_utils::computing_device::ComputingDevice;
 use crate::utils::features::*;
 use crate::measure_time;
