+++ analysis/src/shared_utils/computing_device.rs
