+++ cli/examples/analysis_audio_file.rs
@@ -1,6 +1,6 @@
 use analysis::{
     analysis::{analyze_audio, normalize_analysis_result},
-    shared_utils::computing_device_type::ComputingDevice,
+    shared_utils::computing_device::ComputingDevice,
 };
 
 fn main() {
