+++ cli/examples/build_database.rs
@@ -2,7 +2,7 @@ use log::{error, info};
 use std::path::PathBuf;
 use tracing_subscriber::filter::EnvFilter;
 
-use analysis::shared_utils::computing_device_type::ComputingDevice;
+use analysis::shared_utils::computing_device::ComputingDevice;
 use database::actions::analysis::{
     analysis_audio_library, empty_progress_callback as empty_analysis_progress_callback,
 };
